+++ src/Statuses.php
@@ -2,7 +2,7 @@
 
 namespace Pronamic\WordPress\Pay\Gateways\EMS\ECommerce;
 
-use Pronamic\WordPress\Pay\Core\Statuses as Core_Statuses;
+use Pronamic\WordPress\Pay\Payments\PaymentStatus as Core_Statuses;
 
 /**
  * Title: EMS e-Commerce Gateway statuses constants
@@ -11,7 +11,7 @@ use Pronamic\WordPress\Pay\Core\Statuses as Core_Statuses;
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 2.0.0
+ * @version 2.0.4
  * @since 1.0.0
  */
 class Statuses {